fixup! выполняет Отрисуй меня полностью (часть 2)
