+++ js/ad-form.js
@@ -1,0 +1,17 @@
+const adForm = document.querySelector('.ad-form');
+
+export const deactivateAdForm = () => {
+  adForm.classList.add('ad-form--disabled');
+
+  Array.from(adForm.elements).forEach((element) => {
+    element.disabled = true;
+  });
+};
+
+export const activateAdForm = () => {
+  adForm.classList.remove('ad-form--disabled');
+
+  Array.from(adForm.elements).forEach((element) => {
+    element.disabled = false;
+  });
+};
