+++ js/main.js
@@ -1,7 +1,7 @@
 import {createAdv} from './mocks/adv.js';
 import {createAdCard} from './ad-card.js';
-import {inactivePage} from './forms-state.js';
-import {activePage} from './forms-state.js';
+import {deactivateAdForm} from './ad-form.js';
+import {deactivateMapFiltersForm} from './map-filters-form.js';
 
 const ad = createAdv();
 const adCard = createAdCard(ad);
@@ -10,4 +10,5 @@ const mapCanvas = document.querySelector('.map__canvas');
 
 mapCanvas.appendChild(adCard);
 
-inactivePage();
+deactivateAdForm();
+deactivateMapFiltersForm();
